packages/scenarios/pyproject.toml
[project]
name = "botcheck-scenarios"
version = "0.1.0"
description = "BotCheck Scenario DSL — Pydantic models, rubric defaults, evidence format"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["botcheck_scenarios"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
